Refactoring persistence service
- Moved cpu metrics and scenario into separate files under the
  persistence_service module

- Created generic PersistenceService trait
- Create Local and Remote implementations for cpu metrics and scenarios
  of the PersistenceService trait
- Written tests for Local services
