+++ Cargo.toml
@@ -38,3 +38,4 @@ sqlx = { version = "0.7.4", features = [
   "sqlite",
   "migrate",
 ] }
+reqwest = { version = "0.12.4", features = ["json"] }
