+++ src/lib.rs
@@ -1,6 +1,6 @@
 pub mod clap_args;
-pub mod database;
 pub mod metrics_server;
+pub mod persistence_service;
 pub mod scenario_runner;
 pub mod settings;
 pub mod telegraf;
